+++ sample-android/src/main/java/care/data4life/sample/DocumentDetailsActivity.java
@@ -68,7 +68,7 @@ import care.data4life.fhir.stu3.model.FhirDecimal;
 import care.data4life.fhir.stu3.model.FhirInstant;
 import care.data4life.fhir.stu3.model.FhirTime;
 import care.data4life.sdk.Data4LifeClient;
-import care.data4life.sdk.Task;
+import care.data4life.sdk.call.Task;
 import care.data4life.sdk.config.DataRestrictionException;
 import care.data4life.sdk.helpers.stu3.AttachmentExtension;
 import care.data4life.sdk.helpers.stu3.DocumentReferenceExtension;
