+++ sdk-core/src/main/java/care/data4life/sdk/call/Task.java
@@ -14,7 +14,7 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.sdk;
+package care.data4life.sdk.call;
 
 import javax.annotation.Nullable;
