+++ sdk-core/src/test/java/care/data4life/sdk/call/CallHandlerTest.kt
@@ -13,16 +13,34 @@
  * applications and/or if you’d like to contribute to the development of the SDK, please
  * contact D4L by email to help@data4life.care.
  */
+package care.data4life.sdk.call
 
-package care.data4life.sdk.listener;
+import care.data4life.sdk.SdkContract
+import io.mockk.mockk
+import junit.framework.Assert.assertTrue
+import org.junit.Before
+import org.junit.Ignore
+import org.junit.Test
 
-import care.data4life.sdk.lang.D4LException;
+class CallHandlerTest {
 
-/**
- * Callback is a listener used when there is only success or error.
- */
-public interface Callback {
-    void onSuccess();
+    private lateinit var errorHandler: SdkContract.ErrorHandler
 
-    void onError(D4LException exception);
+
+    // SUT
+    private lateinit var callHandler: CallHandler
+
+
+    @Before
+    fun setup() {
+        errorHandler = mockk()
+
+        callHandler = CallHandler(errorHandler)
+    }
+
+    @Test
+    @Ignore
+    fun test() {
+        assertTrue(false)
+    }
 }
