+++ sdk-jvm/src/main/java/care/data4life/sdk/Data4LifeClient.java
@@ -25,6 +25,7 @@ import care.data4life.auth.AuthorizationContract;
 import care.data4life.auth.AuthorizationException;
 import care.data4life.auth.AuthorizationService;
 import care.data4life.auth.storage.InMemoryAuthStorage;
+import care.data4life.sdk.call.CallHandler;
 import care.data4life.sdk.log.Log;
 import care.data4life.sdk.network.Environment;
 import care.data4life.securestore.SecureStore;
